iter1 fix12

+++ cmd/gophermart/models/models.go
@@ -64,8 +64,8 @@ type Balance struct {
 }
 
 type BalanceResponse struct {
-	Current    int       `json:"current"`
-	Withdrawn  int       `json:"withdrawn"`
+	Current    float64   `json:"current"`
+	Withdrawn  float64   `json:"withdrawn"`
 	UploadedAt time.Time `json:"uploaded_at"`
 }
